chore: refactor usePressed
