+++ packages/ui/src/composables/index.ts
@@ -18,7 +18,6 @@ export * from './std/browser/observer/useIntersectionObserver'
 export * from './std/event/useLongPress'
 export * from './useMaxSelections'
 export * from './useParsableMeasure'
-export * from './usePressed'
 export * from './std/browser/observer/useResizeObserver'
 export * from './useRouterLink'
 export * from './useScroll'
@@ -66,3 +65,4 @@ export * from './std/event/useLongPress'
 export * from './std/utils/useDebounceFn'
 export * from './std/utils/useThrottleFn'
 export * from './std/utils/useThrottleValue'
+export * from './std/browser/useElementPressed'
