+++ packages/ui/src/composables/std/index.ts
@@ -18,6 +18,12 @@ export * from './browser/useElementWidth'
 export * from './browser/useElementScrollableParent'
 export * from './browser/useElementFocused'
 export * from './browser/useElementFocusedWithin'
+export * from './browser/useElementBackground'
+export * from './browser/observer/useIntersectionObserver'
+export * from './browser/observer/useResizeObserver'
+export * from './browser/useElementRect'
+export * from './browser/useElementPressed'
+export * from './browser/useElementFocusedKeyboard'
 export * from './event/useClickOutside'
 export * from './internal/useElementTemplateRef'
 export * from './internal/useSelectorTemplateRef'
